parallelize glm time shifts

+++ Experiments/ucla_human_full.py
@@ -98,103 +98,95 @@ for i, row in events_raw.iterrows():
 # %%
 # Pipeline 1: raw -> power doppler
 
-raw_data_files = (experiment_folder / "raw_frame_data").glob("[!.]*.h5")
-pwd_frames = dict()
-for raw_data_file in tqdm(list(raw_data_files)):
-    match = re.search(r'\d{4}-\d{2}-\d{2}T\d{2}-\d{2}-\d{2}-\d{6}', raw_data_file.stem)
-    timestamp = datetime.strptime(match.group(0).replace('-', ':'), '%Y:%m:%dT%H:%M:%S:%f')
-    # %%
-    # Preprocess
-    data, scan_timing_metadata = preprocess_raw_data(
-        metadata, raw_data_file, transducer_metadata=transducer_metadata
-    )
-    # file = h5py.File(str(raw_data_file).replace('raw_frame_data', 'schema'))
-    # np.testing.assert_array_equal(data, file['data'][:])
+def raw_to_power_doppler(experiment_folder):
+    """Convert raw, bit-packed data to power doppler."""
+    pwd_files = list((experiment_folder / "power_doppler").glob("[!.]*.h5"))
+    if pwd_files:   # pre-computed
+        pwd_frames = dict()
+        for pwd_file in tqdm(pwd_files):
+            match = re.search(r'\d{4}-\d{2}-\d{2}T\d{2}-\d{2}-\d{2}-\d{6}',
+                              pwd_file.stem)
+            timestamp = datetime.strptime(
+                match.group(0).replace('-', ':'), '%Y:%m:%dT%H:%M:%S:%f')
+            pwd_frames[timestamp] = h5py.File(pwd_file)['power_doppler'][:]
+
+    """ # partially pre-computed, don't use too slow
+    bmode_files = list((experiment_folder / "beamformed").glob("[!.]*.h5"))
+    elif bmode_files:
+        pwd_frames = dict()
+        for bmode_file in tqdm(bmode_files):
+            match = re.search(r'\d{4}-\d{2}-\d{2}T\d{2}-\d{2}-\d{2}-\d{6}',
+                              bmode_file.stem)
+            timestamp = datetime.strptime(
+                match.group(0).replace('-', ':'), '%Y:%m:%dT%H:%M:%S:%f')
+            beamformed = h5py.File(bmode_file)['beamformed'][:]
+            pwd_frame = PCAMetalPowerDoppler(
+                num_tissue_components=50,
+                num_blood_and_tissue_components=None
+            ).fit_transform(
+                beamformed.transpose(1, 2, 3, 0)
+            )
+            del beamformed
+            pwd_frames[timestamp] = pwd_frame
     """
-    rf_data = preprocess_iq_data_to_rf(
-        data, iq_afe_artifact_data=None, interp_factor=1,
-        freq_demodulation=metadata.poseidon_config.tx_freq_hz
-    )
-    """
-    # %%
-    # vbeam
-    vbeam_setup = import_space_time_to_vbeam_setup(
-        rf_or_iq_data=data,
-        acquisition_sequence_metadata=acquisition_sequence_metadata,
-        speed_of_sound_m_s=1540.0,
-    )
-    del data
-    # Naive way to reduce GPU-RAM usage
-    # Slice by:
-    # - frame
-    # then beamform each separately
-    vbeam_setup_slice = vbeam_setup.slice["frames", 0]
-    beamformer = das_beamformer(
-        setup=vbeam_setup_slice,
-        ram_limit_gb=8.0,
-    )
-    beamformer = jax.jit(beamformer)
-    beamformed = np.array(
-        _time_beamform(vbeam_setup=vbeam_setup, beamformer=beamformer)
-    )
-    # fails: settings changed in between runs
-    # file = h5py.File(str(raw_data_file).replace('raw_frame_data', 'beamformed'))
-    # np.testing.assert_array_equal(beamformed, file['beamformed'][:])
+    if not pwd_files:
+        raw_data_files = list((experiment_folder / "raw_frame_data").glob("[!.]*.h5"))
+        pwd_frames = dict()
+        for raw_data_file in tqdm(raw_data_files):
+            match = re.search(r'\d{4}-\d{2}-\d{2}T\d{2}-\d{2}-\d{2}-\d{6}', raw_data_file.stem)
+            timestamp = datetime.strptime(match.group(0).replace('-', ':'), '%Y:%m:%dT%H:%M:%S:%f')
+            # %%
+            # Preprocess
+            data, scan_timing_metadata = preprocess_raw_data(
+                metadata, raw_data_file, transducer_metadata=transducer_metadata
+            )
+            # %%
+            # vbeam
+            vbeam_setup = import_space_time_to_vbeam_setup(
+                rf_or_iq_data=data,
+                acquisition_sequence_metadata=acquisition_sequence_metadata,
+                speed_of_sound_m_s=1540.0,
+            )
+            del data
+            vbeam_setup_slice = vbeam_setup.slice["frames", 0]
+            beamformer = das_beamformer(
+                setup=vbeam_setup_slice,
+                ram_limit_gb=8.0,
+            )
+            beamformer = jax.jit(beamformer)
+            beamformed = np.array(
+                _time_beamform(vbeam_setup=vbeam_setup, beamformer=beamformer)
+            )
 
-    # %%
-    # Power doppler
-    pwd_frame = PCAMetalPowerDoppler(
-        num_tissue_components=50,
-        num_blood_and_tissue_components=None
-    ).fit_transform(
-        beamformed.transpose(1, 2, 3, 0)
+            # %%
+            # Power doppler
+            pwd_frame = PCAMetalPowerDoppler(
+                num_tissue_components=50,
+                num_blood_and_tissue_components=None
+            ).fit_transform(
+                beamformed.transpose(1, 2, 3, 0)
+            )
+            del beamformed
+
+            pwd_frames[timestamp] = pwd_frame
+
+    time_stamps = sorted(pwd_frames)
+    pwd = nib.Nifti1Image(
+        np.array([pwd_frames.get(timestamp)
+                  for timestamp in time_stamps]).transpose(1, 2, 3, 0),
+        np.diag([0.15, 0.15, 0.15, 1])
     )
-    del beamformed
-    # Also fails: because of backend? Stochastic?
-    # fname = list((raw_data_file.parent.parent / 'power_doppler').glob(f'{raw_data_file.stem}*.h5'))[0]
-    # assert 'num_blood_and_tissue_components=None' in fname.stem, 'num_tissue_components=50' in fname.stem
-    # np.testing.assert_array_equal(pwd_frame, h5py.File(fname)['power_doppler'][:])
+    del pwd_frames
+    time_stamps = np.array([(timestamp - time_stamps[0]).total_seconds()
+                            for timestamp in time_stamps])
+    return pwd, time_stamps
 
-    pwd_frames[timestamp] = pwd_frame
 
-""" # for use if bmode is done but not pwd
-bmode_files = list((experiment_folder / "beamformed").glob("[!.]*.h5"))
-pwd_frames = dict()
-for bmode_file in tqdm(bmode_files):
-    match = re.search(r'\d{4}-\d{2}-\d{2}T\d{2}-\d{2}-\d{2}-\d{6}', bmode_file.stem)
-    timestamp = datetime.strptime(match.group(0).replace('-', ':'), '%Y:%m:%dT%H:%M:%S:%f')
-    beamformed = h5py.File(bmode_file)['beamformed'][:]
-    pwd_frame = PCAMetalPowerDoppler(
-        num_tissue_components=50,
-        num_blood_and_tissue_components=None
-    ).fit_transform(
-        beamformed.transpose(1, 2, 3, 0)
-    )
-    del beamformed
-    pwd_frames[timestamp] = pwd_frame
-"""
-
-"""  # for use if bmode and power doppler are done
-pwd_files = (experiment_folder / "power_doppler").glob("[!.]*.h5")
-pwd_frames = dict()
-for pwd_file in tqdm(list(pwd_files)):
-    match = re.search(r'\d{4}-\d{2}-\d{2}T\d{2}-\d{2}-\d{2}-\d{6}', pwd_file.stem)
-    timestamp = datetime.strptime(match.group(0).replace('-', ':'), '%Y:%m:%dT%H:%M:%S:%f')
-    pwd_frames[timestamp] = h5py.File(pwd_file)['power_doppler'][:]
-"""
-
-time_stamps = sorted(pwd_frames)
-pwd = nib.Nifti1Image(
-    np.array([pwd_frames.get(timestamp)
-              for timestamp in time_stamps]).transpose(1, 2, 3, 0),
-    np.diag([0.15, 0.15, 0.15, 1])
-)
 (experiment_folder / 'power_doppler').mkdir(parents=True, exist_ok=True)
-nib.save(pwd, experiment_folder / 'power_doppler' / 'pwd.nii.gz')
-del pwd_frames
-time_stamps = np.array([(timestamp - time_stamps[0]).total_seconds()
-                        for timestamp in time_stamps])
-np.savetxt(experiment_folder / 'power_doppler' / 'pwd_timestamps.txt', time_stamps)
+if not (experiment_folder / 'power_doppler' / 'pwd.nii.gz').exists():
+    pwd, time_stamps = raw_to_power_doppler(experiment_folder)
+    nib.save(pwd, experiment_folder / 'power_doppler' / 'pwd.nii.gz')
+    np.savetxt(experiment_folder / 'power_doppler' / 'pwd_timestamps.txt', time_stamps)
 
 # %%
 # Pipeline 2: power doppler -> zmap
@@ -255,23 +247,26 @@ else:
     np.savetxt(experiment_folder / 'reg' / 'transformations.txt', transformations)
 
 
-# glm
-(experiment_folder / 'glm').mkdir(parents=True, exist_ok=True)
+# define events of interest
 event = min(set(events.trial_type))
 events2 = events[events['trial_type'] == event]
-shifts, max_tstats, best_offset = fit_glm_time_shift(
-    pwd_reg, time_stamps,
-    event, events2, shift=20,
-    smoothing_fwhm=smoothing_fwhm, hrf='glover',
-    transformations=transformations,
-    out_dir=experiment_folder / 'glm'
-)
-fig, ax = plt.subplots()
-ax.plot(shifts, max_tstats)
-ax.set_xlabel('shifts')
-ax.set_ylabel('max_z')
-fig.savefig(experiment_folder / 'glm' / 'shifts.png')
-plt.close(fig)
+
+# glm
+(experiment_folder / 'glm').mkdir(parents=True, exist_ok=True)
+if not (experiment_folder / 'glm' / 'shifts.png').exists():
+    shifts, max_tstats, best_offset = fit_glm_time_shift(
+        pwd_reg, time_stamps,
+        event, events2, shift=20,
+        smoothing_fwhm=smoothing_fwhm, hrf='glover',
+        transformations=transformations,
+        out_dir=experiment_folder / 'glm'
+    )
+    fig, ax = plt.subplots()
+    ax.plot(shifts, max_tstats)
+    ax.set_xlabel('shifts')
+    ax.set_ylabel('max_z')
+    fig.savefig(experiment_folder / 'glm' / 'shifts.png')
+    plt.close(fig)
 
 
 """pwd, time_stamps = get_power_doppler_nii(
